Spawn Point & Checkpoint
Created and implemented a spawn point and checkpoint. Player teleports
to spawn point at beginning of level when its Start() is called. When
player passes checkpoint, static position stored in SpawnPointScript =
location of checkpoint. When the player dies (in player death script),
it calls the SpawnPointScript to teleport the player to the last
checkpoint.